[FEATURE] The null explicit has been added

+++ RSMS.App/RSMS.Entities/DTOs/Properties/CreatePropertyDTO.cs
@@ -1,12 +1,12 @@
 ﻿namespace RSMS.Entities.DTOs.Properties;
-public class CreatePropertyDTO(short propertyTypeId, int ownerId,
-                               string number, string address, float area,
-                               float? constructionArea)
+public class CreatePropertyDTO(byte propertyTypeId, int ownerId,
+                               string number, string address, decimal area,
+                               decimal? constructionArea)
 {
-    public short PropertyTypeId => propertyTypeId;
+    public byte PropertyTypeId => propertyTypeId;
     public int OwnerId => ownerId;
     public string Number => number;
     public string Address => address;
-    public float Area => area;
-    public float? ConstructionArea => constructionArea;
+    public decimal Area => area;
+    public decimal? ConstructionArea => constructionArea;
 }
